Work on sql planning example
